# Optimize shape rendering (pen creation mostly).