## What does the pull request do?
<!--- Give a bit of background on the PR here, together with links to with related issues etc. -->
An optimization for core `Shape` rendering that minimizes perf and memory impact.
Best case (no brushes) improved cpu perf x200 and removed all allocations.
Worst case we are looking at cpu perf x38 and 98% memory reduction.

## What is the current behavior?
<!--- If the PR is a fix, describe the current incorrect behavior, otherwise delete this section. -->
|                        Method |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------:|----------:|----------:|-------:|-------:|------:|----------:|
|         Render_Line_NoBrushes |  9.246 us | 0.0540 us | 0.0506 us | 0.7019 | 0.0153 |     - |   3.01 KB |
|        Render_Line_WithStroke | 14.182 us | 0.0617 us | 0.0577 us | 0.7782 | 0.0305 |     - |   3.32 KB |
|          Render_Line_WithFill |  9.320 us | 0.0447 us | 0.0373 us | 0.7019 | 0.0153 |     - |   3.01 KB |
| Render_Line_WithFillAndStroke | 13.953 us | 0.0374 us | 0.0312 us | 0.7782 | 0.0305 |     - |   3.32 KB |

## What is the updated/expected behavior with this PR?
<!--- Describe how to test the PR. -->
|                        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |----------:|---------:|---------:|-------:|------:|------:|----------:|
|         Render_Line_NoBrushes |  45.30 ns | 0.092 ns | 0.086 ns |      - |     - |     - |         - |
|        Render_Line_WithStroke | 339.26 ns | 1.937 ns | 1.812 ns | 0.0210 |     - |     - |      88 B |
|          Render_Line_WithFill |  87.35 ns | 0.374 ns | 0.332 ns |      - |     - |     - |         - |
| Render_Line_WithFillAndStroke | 352.55 ns | 2.207 ns | 2.065 ns | 0.0210 |     - |     - |      88 B |

## How was the solution implemented (if it's not obvious)?
<!--- Include any information that might be of use to a reviewer here. -->
Best optimization is to just avoid doing any work, so when `Stroke` is not set we can just ignore it.
Also majority of rendered shapes are not using dashes so there is no point in allocating dash style for everything.

Biggest change is using immutable pen and dash styles. Mutable variants have an absurd overhead due to change tracking.

## Checklist

- [ ] Added unit tests (if possible)?
- [ ] Added XML documentation to any related classes?
- [ ] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Breaking changes
<!--- List any breaking changes here. When the PR is merged please add an entry to https://github.com/AvaloniaUI/Avalonia/wiki/Breaking-Changes -->

## Obsoletions / Deprecations
<!--- Obsolete and Deprecated attributes on APIs MUST only be included when discussed with Core team. @grokys, @kekekeks & @danwalmsley -->

## Fixed issues
<!--- If the pull request fixes issue(s) list them like this: 
Fixes #123
Fixes #456
-->